producers/tmdb/requirements.txt
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
colorlog==6.8.0
schedule==1.2.0
